Now using a helper for the tests
