+++ test/mocha.opts
@@ -1,1 +1,3 @@
 --compilers js:babel-core/register
+--require ./test/helper.js
+--check-leaks
