+++ test/test_createAgent.js
@@ -1,18 +1,4 @@
-import _ from 'lodash';
 import craftai, { STATUS } from '../src';
-import dotenv from 'dotenv';
-import { expect } from 'chai';
-
-dotenv.config({silent: true});
-dotenv.load();
-
-const CRAFT_CFG = {
-  owner: 'craft-ai',
-  name: 'craft-ai-client-js',
-  version: 'master',
-  appId: process.env.CRAFT_APP_ID,
-  appSecret: process.env.CRAFT_APP_SECRET
-}
 
 describe('instance', function() {
   this.timeout(5000);
